pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.0.1"
description = ""
authors = ["Gabriel Aranha <gabriel.aranha@makes.ai>"]

[tool.poetry.dependencies]
python = ">=3.8, <3.10"
fastapi = "0.75.0"
uvicorn = {extras = ["standard"], version = "0.17.6"}
python-decouple = "3.6"
fastapi-pagination = "0.9.3"
python-json-logger = "2.0.4"
alembic = "1.8.1"
aioredis = "2.0.1"
requests = "^2.28.1"
pymysql = "1.0.2"

[tool.poetry.dev-dependencies]  # Deprecated, use -G dev instead
pytest = "7.1.2"
coverage = "6.4.4"
pytest-mock = "3.8.2"
pytest-asyncio = "0.19.0"
doppler-env = "^0.3.0"
pytest-cov = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.3"
pylint-pydantic = "^0.1.4"
mypy = "^0.982"
sqlalchemy2-stubs = "^0.0.2a27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
